requirements.txt
# CVE Processing Azure Function Dependencies

# Azure Functions
azure-functions>=1.18.0

# Azure SQL Database connectivity
pyodbc>=4.0.35

# NVD API integration
nvdlib>=0.7.3

# AI analysis
openai>=1.3.0

# Web scraping and content processing
requests>=2.31.0
beautifulsoup4>=4.12.0
readability-lxml>=0.8.1

# Token counting for AI
tiktoken>=0.5.0

# Additional utilities
python-dateutil>=2.8.2